move ResetRestart under main
