+++ src/main.py
@@ -3,7 +3,7 @@ import sys
 from src.helper.main.converter.converter import converter_exit, converter_string_list_int, converter_string_boolean, \
     converter_string_int
 from src.helper.main.main import run_method_with_nodes, checker, show_help, select_models, select_method
-from src.network.ResetRestart import check_reset_restart
+from src.helper.main.ResetRestart import check_reset_restart
 from src.network.StarNetwork import StarNetwork
 from src.protocols.Experiment import Experiment
